feature: correct hash merge lookup from hiera
By moving the hiera_hash function into the main function, it's now.
possible to have a correct hash merge lookup. This commit also brings
the pssibility to configure the Hiera key names and the manage_users
and manage_groups parameters to be a bit more flexible.
It also contains updates to the RSpec tests to include Hiera in the.
tests.