Invoke pgrep/pkill with symbolical username
Commit 8fbc25a added code to forcefully terminate processes before
removing user accounts. The code made the assumption that UIDs are not
reused and used the numeric ID to terminate processes. The basis for the
assumption was that the mapping from username to ID could already have
been removed while there are still processes running. Those processes
could be selected using the numeric user ID.

Unfortunately there has been at least one case where a numeric user ID
was reused, leading to processes being terminated on every Puppet run.
The original user was removed a few months ago and now the ID is being
used for a service.

With this change pgrep/pkill are always invoked using the symbolical
username.